refactor: eliminate poor abstraction in hrm-authentictopackage
