+++ lambdas/packages/hrm-authentication/index.ts
@@ -14,25 +14,6 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 import { TResult } from '@tech-matters/types';
-import filesUrlsAuthenticator, {
-  HrmAuthenticateFilesUrlsRequestData,
-} from './filesUrlsAuthenticator';
-
-/**
- * The authenticator will call the authenticator based on the type.
- * In a perfect world the hrm side of authentication would be a single endpoint
- * that would accept a common payload and return a common response.
- * And this very leaky abstraction would not be needed.
- *
- * For now we have to support multiple endpoints and multiple payloads with
- * different responses, so the function is basically an adapter.
- *
- * The goal was to keep all hrm authentication transformations centralized
- * in a single place to aid in the future refactoring.
- */
-const types = {
-  filesUrls: (params: HrmAuthenticateParameters) => filesUrlsAuthenticator(params),
-};
 
 /**
  * The object types that can be authenticated.
@@ -48,23 +29,6 @@ export const isAuthenticationObjectType = (
   type: string,
 ): type is HRMAuthenticationObjectTypes => Object.keys(objectTypes).includes(type);
 
-export type HrmAuthenticateTypes = keyof typeof types;
-
 export type HrmAuthenticateResult = TResult<true>;
 
-export type HrmAuthenticateParameters = {
-  accountSid: string;
-  objectType: HRMAuthenticationObjectTypes;
-  objectId?: string;
-  type: HrmAuthenticateTypes;
-  authHeader: string;
-  requestData: HrmAuthenticateFilesUrlsRequestData;
-};
-
-export const authenticate = async (
-  params: HrmAuthenticateParameters,
-): Promise<HrmAuthenticateResult> => {
-  return types[params.type](params);
-};
-
-export { FileTypes } from './filesUrlsAuthenticator';
+export * from './callHrmApi';
